@@ -16,24 +16,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "10.0.4",
-    "@angular/common": "10.0.4",
-    "@angular/compiler": "10.0.4",
-    "@angular/core": "10.0.4",
-    "@angular/platform-browser": "10.0.4",
-    "@angular/platform-browser-dynamic": "10.0.4",
+    "@angular/animations": "11.0.2",
+    "@angular/common": "11.0.2",
+    "@angular/compiler": "11.0.2",
+    "@angular/core": "11.0.2",
+    "@angular/platform-browser": "11.0.2",
+    "@angular/platform-browser-dynamic": "11.0.2",
     "rxjs": "6.5.5",
     "tslib": "^2.0.0",
     "zone.js": "0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1000.3",
-    "@angular/cli": "10.0.3",
-    "@angular/compiler-cli": "10.0.4",
-    "@angular/language-service": "10.0.4",
+    "@angular-devkit/build-angular": "0.1100.2",
+    "@angular/cli": "11.0.2",
+    "@angular/compiler-cli": "11.0.2",
+    "@angular/language-service": "11.0.2",
     "@types/node": "13.13.0",
     "ts-node": "8.8.2",
     "tslint": "~6.1.0",
-    "typescript": "3.9.7"
+    "typescript": "4.0.5"
   }
 }
