@@ -20,11 +20,11 @@
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
   "devDependencies": {
+    "@rollup/plugin-commonjs": "13.0.0",
+    "@rollup/plugin-node-resolve": "8.0.1",
     "rollup": "2.13.1",
-    "@rollup/plugin-commonjs": "^13.0.0",
-    "@rollup/plugin-node-resolve": "^8.0.1",
     "rollup-plugin-svelte": "5.2.2",
-    "rollup-plugin-terser": "^6.1.0",
-    "svelte": "3.28.0"
+    "rollup-plugin-terser": "6.1.0",
+    "svelte": "3.29.4"
   }
 }
