@@ -1,5 +1,5 @@
 {
-  "name": "js-framework-benchmark-vue",
+  "name": "js-framework-benchmark-vue-next",
   "version": "1.0.0",
   "description": "Benchmark for vue.js framework",
   "js-framework-benchmark": {
@@ -20,16 +20,12 @@
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
   "devDependencies": {
-    "@babel/core": "7.2.2",
-    "@babel/preset-env": "7.3.1",
-    "css-loader": "2.1.0",
-    "vue-loader": "15.6.2",
-    "babel-loader": "8.0.5",
-    "vue-template-compiler": "2.6.2",
-    "webpack": "4.29.2",
-    "webpack-cli": "3.2.3"
+    "@vue/compiler-sfc": "3.0.2",
+    "vue-loader": "16.0.0-rc.2",
+    "webpack": "4.41.4",
+    "webpack-cli": "3.3.10"
   },
   "dependencies": {
-    "vue": "2.6.2"
+    "vue": "3.0.2"
   }
 }
