@@ -24,29 +24,29 @@
   "author": "",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@types/lodash": "4.14.161",
-    "@types/node": "14.6.2",
-    "@types/ramda": "0.27.14",
-    "@types/selenium-webdriver": "4.0.9",
-    "@types/semver": "7.3.3",
-    "@types/yargs": "15.0.5",
+    "@types/lodash": "4.14.165",
+    "@types/node": "14.14.10",
+    "@types/ramda": "0.27.32",
+    "@types/selenium-webdriver": "4.0.10",
+    "@types/semver": "7.3.4",
+    "@types/yargs": "15.0.10",
     "dockerode": "^3.2.1",
     "rimraf": "^3.0.2",
     "ts-node": "9.0.0",
-    "typescript": "4.0.2"
+    "typescript": "4.1.2"
   },
   "dependencies": {
-    "axios": "^0.20.0",
-    "chromedriver": "^85.0.0",
+    "axios": "^0.21.0",
+    "chromedriver": "^87.0.0",
     "cross-env": "7.0.2",
     "dot": "1.1.3",
-    "jstat": "1.9.3",
-    "lighthouse": "6.3.0",
+    "jstat": "1.9.4",
+    "lighthouse": "6.4.1",
     "lodash": "4.17.20",
-    "npm-check-updates": "8.0.1",
+    "npm-check-updates": "10.2.2",
     "ramda": "0.27.1",
     "selenium-webdriver": "4.0.0-alpha.7",
     "semver": "7.3.2",
-    "yargs": "15.4.1"
+    "yargs": "16.1.1"
   }
 }
